subcoin.sol version - 2
Here we are keeping this mint function to work only once and minted by the minter only... We closed all the loops using modifiers... Also, we guarantee the investor's investment protection for 2 years... Though we are using block. timestamp, which can be a little dangerous... But it does the job for now.